refactor(groupbox): change GroupBox to AXAML
